Refactor main.rb in order to respect SOLID principle
